Improve security and front end interaction

+++ src/main/java/com/example/backend/Controllers/PollController.java
@@ -1,7 +1,14 @@
 package com.example.backend.Controllers;
 
 import java.util.HashMap;
+import java.util.List;
 
+import org.h2.engine.Role;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,11 +16,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.backend.Managers.PollManager;
 import com.example.backend.Model.Poll.Poll;
 import com.example.backend.Model.Poll.PollRequest;
+import com.example.backend.Model.User.User;
 
 @RestController
 @CrossOrigin
@@ -34,9 +43,17 @@ public class PollController {
     }
 
     @DeleteMapping("/{pollID}")
-    public boolean deletePoll(@PathVariable("pollID") Integer pollID) {
-        return manager.deletePoll(pollID);
-
+    public ResponseEntity<?> deletePoll(@PathVariable Integer pollID, @AuthenticationPrincipal UserDetails user) {
+        var pollOpt = manager.getPoll(pollID);
+        if (pollOpt.isEmpty()) {
+            return ResponseEntity.notFound().build();
+        }
+        var poll = pollOpt.get();
+        if (!poll.getCreatedBy().getUsername().equals(user.getUsername())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+        manager.deletePoll(pollID);
+        return ResponseEntity.noContent().build();
     }
 
     @GetMapping("/{pollID}")
@@ -47,6 +64,22 @@ public class PollController {
             throw e;
         }
 
+
+        
     }
+
+    @GetMapping
+    public ResponseEntity<List<Poll>> getPolls(
+                @RequestParam(value = "user", required = false) String username) {
+            List<Poll> polls;
+            if (username != null) {
+                polls = manager.getPollsByUserName(username);
+            } else {
+                polls = manager.getPolls();
+            }
+            return ResponseEntity.ok(polls);
+        }
+
+
 
 }
